requirements-dev.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements-dev.in
#
async-generator==1.10
    # via
    #   -c requirements.txt
    #   trio
    #   trio-websocket
attrs==21.2.0
    # via
    #   -c requirements.txt
    #   outcome
    #   pytest
    #   trio
autoflake==1.4
    # via -r requirements-dev.in
autopep8==1.6.0
    # via -r requirements-dev.in
black==21.12b0
    # via -r requirements-dev.in
certifi==2021.10.8
    # via
    #   -c requirements.txt
    #   requests
    #   urllib3
cffi==1.15.0
    # via
    #   -c requirements.txt
    #   cryptography
charset-normalizer==2.0.7
    # via
    #   -c requirements.txt
    #   requests
click==7.1.2
    # via
    #   -c requirements.txt
    #   black
coverage[toml]==6.2
    # via
    #   -r requirements-dev.in
    #   pytest-cov
cryptography==35.0.0
    # via
    #   -c requirements.txt
    #   pyopenssl
    #   urllib3
django-coverage==1.2.4
    # via -r requirements-dev.in
freezegun==1.1.0
    # via -r requirements-dev.in
h11==0.12.0
    # via
    #   -c requirements.txt
    #   wsproto
idna==3.3
    # via
    #   -c requirements.txt
    #   requests
    #   trio
    #   urllib3
iniconfig==1.1.1
    # via
    #   -c requirements.txt
    #   pytest
isort==5.10.1
    # via -r requirements-dev.in
mccabe==0.6.1
    # via pylama
mypy-extensions==0.4.3
    # via black
outcome==1.1.0
    # via
    #   -c requirements.txt
    #   trio
packaging==21.2
    # via
    #   -c requirements.txt
    #   pytest
pathspec==0.9.0
    # via black
platformdirs==2.4.1
    # via black
pluggy==1.0.0
    # via
    #   -c requirements.txt
    #   pytest
py==1.11.0
    # via
    #   -c requirements.txt
    #   pytest
pycodestyle==2.8.0
    # via
    #   autopep8
    #   pylama
pycparser==2.21
    # via
    #   -c requirements.txt
    #   cffi
pydocstyle==6.1.1
    # via pylama
pyflakes==2.4.0
    # via
    #   autoflake
    #   pylama
pylama==8.2.1
    # via -r requirements-dev.in
pyopenssl==21.0.0
    # via
    #   -c requirements.txt
    #   urllib3
pyparsing==2.4.7
    # via
    #   -c requirements.txt
    #   packaging
pytest==6.2.5
    # via
    #   -c requirements.txt
    #   -r requirements-dev.in
    #   pytest-base-url
    #   pytest-cov
    #   pytest-django
    #   pytest-html
    #   pytest-logging
    #   pytest-metadata
    #   pytest-mock
    #   pytest-responses
    #   pytest-selenium
    #   pytest-variables
pytest-base-url==1.4.2
    # via pytest-selenium
pytest-cov==3.0.0
    # via -r requirements-dev.in
pytest-django==4.4.0
    # via -r requirements-dev.in
pytest-html==3.1.1
    # via
    #   -r requirements-dev.in
    #   pytest-selenium
pytest-logging==2015.11.4
    # via -r requirements-dev.in
pytest-metadata==1.11.0
    # via pytest-html
pytest-mock==3.6.1
    # via -r requirements-dev.in
pytest-responses==0.5.0
    # via -r requirements-dev.in
pytest-selenium==2.0.1
    # via -r requirements-dev.in
pytest-variables==1.9.0
    # via pytest-selenium
python-dateutil==2.8.2
    # via freezegun
requests==2.26.0
    # via
    #   -c requirements.txt
    #   pytest-base-url
    #   pytest-selenium
    #   requests-mock
    #   responses
requests-mock==1.9.3
    # via -r requirements-dev.in
responses==0.16.0
    # via pytest-responses
selenium==4.0.0
    # via
    #   -c requirements.txt
    #   pytest-selenium
six==1.16.0
    # via
    #   -c requirements.txt
    #   pyopenssl
    #   python-dateutil
    #   requests-mock
    #   responses
    #   tenacity
sniffio==1.2.0
    # via
    #   -c requirements.txt
    #   trio
snowballstemmer==2.2.0
    # via pydocstyle
sortedcontainers==2.4.0
    # via
    #   -c requirements.txt
    #   trio
tenacity==6.3.1
    # via pytest-selenium
toml==0.10.2
    # via
    #   -c requirements.txt
    #   autopep8
    #   pytest
tomli==1.2.2
    # via
    #   black
    #   coverage
trio==0.19.0
    # via
    #   -c requirements.txt
    #   selenium
    #   trio-websocket
trio-websocket==0.9.2
    # via
    #   -c requirements.txt
    #   selenium
typing-extensions==4.0.0
    # via
    #   -c requirements.txt
    #   black
urllib3[secure]==1.26.7
    # via
    #   -c requirements.txt
    #   requests
    #   responses
    #   selenium
wsproto==1.0.0
    # via
    #   -c requirements.txt
    #   trio-websocket
